requirements.txt
PyQt5
pystray
Pillow
